[kotlin] fix `KtInaccessibleLifetimeOwnerAccessException` from K2 UAST
`KotlinUFunctionCallExpression.multiresolve()` returns a lazily-populated sequence.

Internally, it works via `FirKotlinUastResolveProviderService.getReferenceVariants`.
This method implementation consists of two parts: collecting resolve candidates and creating a `PsiElement` sequence of those elements.
The second part is lazy by the nature of `Sequence`, so it can be accessed later, possibly in a different read action.
We cannot leak `KtLifetimeOwner` from one `analyze` call to another. So we pass `KtPsiElementPointer` instead of a list of `KtCallCandidateInfo`.

^KT-60539 fixed
^KT-60530 fixed

GitOrigin-RevId: 1e1a02149021328bbc06658b1f74de1aba7a5020