[kotlin] k2: move statement test (KTIJ-26153)
Disabled tests for property accessors.
Move is performed via document, which is translated to psi via commitDocumentThread.
This generates different psi events which might correspond to the psi which can never be parsed, e.g. `KotlinFirMoveStatementTestGenerated.ClassBodyDeclarations.Accessors.testAccessor3` provides a case with a property which has 2 setters, one of which has getter body. Such "setter" can't be found in FIR, because fir checks `isGetter` first.

GitOrigin-RevId: 6e862084af4a49f4464f359c6831f70a74eb08bd